% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_filter.R
\name{causal_filter}
\alias{causal_filter}
\title{Causal Filtering}
\usage{
causal_filter(data, causal_probs, iterations)
}
\arguments{
\item{data}{Data.}

\item{causal_probs}{Causal probabilities, or 1 - p-values from a test for causality, for each interactor (column in the data).}

\item{iteractions}{Number of times to apply the causal filter.}
}
\value{
Values in the Jacobian.
}
\description{
Filters data with supplied probabilities of causality. This can serve a priori in place of post-hoc model selection.
}
\examples{

}
