% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_causalfilter.R
\name{lm_causalfilter}
\alias{lm_causalfilter}
\title{Causal Filtering}
\usage{
lm_causalfilter(x, causal_probs, weights)
}
\arguments{
\item{x}{Time series at time t.}

\item{causal_probs}{Causal probabilities.}

\item{weights}{Weights for the exponentially-weighted linear model.}

\item{iteractions}{Number of times to apply the causal filter.}
}
\value{
Values in the Jacobian.
}
\description{
Applies supplied probabilities of causality to time series data.
}
\examples{

}
