% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_svd.R
\name{lm_svd}
\alias{lm_svd}
\title{SVD solution to the Linear Least Squares Problem}
\usage{
lm_svd(y, x, weights, intercept = TRUE)
}
\arguments{
\item{y}{Target time series at time t+1.}

\item{x}{Time series at time t.}

\item{weights}{Weights for the exponentially-weighted linear model.}
}
\value{
Values in the Jacobian.
}
\description{
Solution to the linear least squares problem using Singular Value Decomposition (SVD).
}
\examples{
x <- matrix(sample(1:10, size = 30, replace = TRUE), nrow = 10, ncol = 3)
y <- 2*x[,1] + 0.25*x[,2] + rnorm(n = 10, mean = 0, sd = 1)
weights <- apply(x, MARGIN = 1, FUN = var)

lm_svd(y, x, weights)

}
\references{
Deyle, E. R., May, R. M., Munch, S. B., & Sugihara, G. (2016, January). Tracking and forecasting ecosystem interactions in real time. In Proc. R. Soc. B (Vol. 283, No. 1822, p. 20152258). The Royal Society.
}
