% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_svdsolve.R
\name{lm_svdsolve}
\alias{lm_svdsolve}
\title{SVD Solution to the Linear Least Squares Problem}
\usage{
lm_svdsolve(y, x, weights)
}
\arguments{
\item{y}{Target time series at time t+1.}

\item{x}{Time series at time t.}

\item{weights}{Weights for the exponentially-weighted linear model.}
}
\value{
Values in the Jacobian.
}
\description{
Solution of the linear least squares problem using Singular Value Decomposition (SVD).
}
\examples{
lm_svdsolve(block[libs, 1], block[libs, 2:dim(block)[2]], weights)

}
\references{
Deyle, E. R., May, R. M., Munch, S. B., & Sugihara, G. (2016, January). Tracking and forecasting ecosystem interactions in real time. In Proc. R. Soc. B (Vol. 283, No. 1822, p. 20152258). The Royal Society.
}
