% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Tracking and Forecasting Interactions in Real Time}
\usage{
track(time_series, target, dates, theta = 8, block = FALSE)
}
\arguments{
\item{time_series}{A data frame where each column is a time series.}

\item{target}{Which time series (column) to calculate the sequential interactions of.}

\item{dates}{Optional vector of dates.}

\item{theta}{Parameter to tune the relationship between distance and weight in the linear model. Defaults to 8 which was used by Deyle, May, Munch, and Sugihara.}

\item{block}{Defaults to FALSE. If your data is stacked spatial replicates, where you have already made the first column for time 1+1 and all others for time t.}
}
\value{
A data frame where each column is a row of the sequential Jacobian.
}
\description{
Calculates sequential Jacobians on time series data using the S-map method developed and detailed in the 2016 Proceedings of the Royal Society B paper titled Tracking and Forecasting Ecological Interactions in Real Time by Deyle, May, Munch, and Sugihara.
}
\examples{
data <- matrix(runif(300, 0, 1), nrow = 100, ncol=3))

track(data, target = 1)

}
\references{
Deyle, E. R., May, R. M., Munch, S. B., & Sugihara, G. (2016, January). Tracking and forecasting ecosystem interactions in real time. In Proc. R. Soc. B (Vol. 283, No. 1822, p. 20152258). The Royal Society.
}
